Defer StartOS Context change to PendSV_Handler
StartOS is executing within an SVC call. Therefore, priority exception
will be set to maximum and if it had to return to a different exception
handler after the SVC call finishes (i.e.: is in a nested interrupt)
the code would crash because we are overwriting the return code in the
link register.

It is only safe to do so in the PendSV_Handler, which has minimum
priority and cannot be in a nested interrupt (avoiding the situation
entirely).

Also, refactor the code for the task initialization to use local
variables in main (that are not freed like before, and therefore don't
need to be static).

Ref. #16